// client/modules/CharacterEditor/Service/Types/FaceType.ts
export type FaceType = {
    noseWidth: number // -1..1
    noseHeight: number // -1..1 invert
    noseLength: number // -1..1 invert
    noseBridge: number // -1..1
    noseTip: number // -1..1 invert
    noseBridgeShift: number // -1..1 invert
    browHeight: number // -1..1 invert
    browWidth: number // -1..1
    jawWidth: number // -1..1
    jawHeight: number // -1..1 invert
    chinLength: number // -1..1
    chinPosition: number // -1..1
    chinWidth: number // -1..1
    chinShape: number // -1..1
    lips: number // -1..1 invert
    cheeksWidth: number // -1..1 invert
    cheekboneHeight: number // -1..1 invert
    cheekboneWidth: number // -1..1
    eyes: number // -1..1 invert
}